Docs: Formatting and more content

+++ include/TransformComponent.h
@@ -5,8 +5,7 @@
 #include "Constants.h"
 #include "DataComponent.h"
 
-//! \brief Adds a transform to an entity when added via entity.addComponent()
-//! \details The transform component is responsible for the position, direction and size of an entity. It is used to move the entity in the game world and to determine its size and position on the screen.
+//! \brief The transform component is responsible for the position, direction and size of an entity. It is used to move the entity in the game world and to determine its size and position on the screen.
 class TransformComponent : public Component
 {
 public:
@@ -15,12 +14,24 @@ public:
 
 	int height = 32;
 	int width = 32;
-	int scale = 1;
+	float scale = 1;
 
 	//! \attention in order to allow an entity to move the stat "speed" must be set in the DataComponent (written exactly like that and set to any positive int value, 0 will lead to no movement, negative numbers to backwards movement)
-	explicit TransformComponent(int scale = 1);
-	TransformComponent(float x, float y, int scale = 1);
-	TransformComponent(float x, float y, int w, int h, int scale = 1);
+
+	//! \param scale base value is 1 (32x32px), size gets multiplied with scale
+	explicit TransformComponent(float scale = 1);
+
+	//! \param x x coordinate of spawnposition
+	//! \param y y coordinate of spawnposition
+	//! \param scale base value is 1 (32x32px per default), size gets multiplied with scale
+	TransformComponent(float x, float y, float scale = 1);
+
+	//! \param x x coordinate of spawnposition
+	//! \param y y coordinate of spawnposition
+	//! \param w add custom width
+	//! \param h add custom height
+	//! \param scale base value is 1 (32x32px per default), size gets multiplied with scale
+	TransformComponent(float x, float y, int w, int h, float scale = 1);
 
 	void init() override;
 	void update(uint_fast16_t diffTime) override;
